+++ src/components/shared/text/Text.jsx
@@ -1,13 +1,7 @@
-import styled, { css } from 'styled-components';
+import StyledText from 'components/shared/text/style';
 
-const StyledText = styled.p`
-  ${props =>
-    props.modal &&
-    css`
-      font-weight: 500;
-      font-size: 25px;
-      color: var(--primary-color);
-    `}
-`;
+const Text = ({ children, ...rest }) => {
+  return <StyledText {...rest}>{children}</StyledText>;
+};
 
-export default StyledText;
+export default Text;
