+++ src/components/shared/text/style.js
@@ -1,7 +1,13 @@
-import StyledText from 'components/shared/text/style';
+import styled, { css } from 'styled-components';
 
-const Text = ({ children, ...rest }) => {
-  return <StyledText {...rest}>{children}</StyledText>;
-};
+const StyledText = styled.p`
+  ${props =>
+    props.modal &&
+    css`
+      font-weight: 500;
+      font-size: 25px;
+      color: var(--primary-color);
+    `}
+`;
 
-export default Text;
+export default StyledText;
