+++ src/page/productList/ProductListPage.jsx
@@ -1,34 +1,15 @@
 import { useSelector } from 'react-redux';
 import { useEffect } from 'react';
-import styled from 'styled-components';
 import axios from 'axios';
 
 import ProductItem from 'components/productItem/ProductItem';
 
 import store from 'store/store';
 
+import { StyledProductListPage, StyledProductList } from 'page/productList/style';
+
 import { INITIALIZE } from 'actions/action';
 import { SERVER_URL } from 'constants';
-
-const StyledProductListPage = styled.main`
-  width: 1269px;
-  background-color: white;
-  margin-top: 140px;
-  padding: 40px;
-  overflow: scroll;
-  height: 100%;
-
-  -ms-overflow-style: none;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-const StyledProductList = styled.div`
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 47px;
-`;
 
 const ProductListPage = () => {
   const { products, shoppingCart } = useSelector(state => state.reducer);
